Add double fault handling, timer PIC handling, deadlock avoidance.
This commit adds double fault handling, including switching to a new
stack when a double fault is triggered. This commit also adds the
beginnings of PIC interrupt handling, printing a dot when the timer
interrupt is received. Finally, this commit adds a bit of deadlock
avoidance for printing, disabling interrupts while printing is going on.